Made the create bot endpoint secure
