+++ modules/discord/app/com/cobble/bot/discord/DiscordBot.scala
@@ -1,5 +1,7 @@
 package com.cobble.bot.discord
 
+import java.net.URLEncoder
+import java.nio.charset.StandardCharsets
 import javax.inject.Singleton
 
 import com.google.inject.Inject
@@ -7,6 +9,8 @@ import play.api.Configuration
 import play.api.inject.ApplicationLifecycle
 import sx.blah.discord.api.events.EventDispatcher
 import sx.blah.discord.api.{ClientBuilder, IDiscordClient}
+import sx.blah.discord.handle.obj.Permissions
+import sx.blah.discord.util.BotInviteBuilder
 
 import scala.concurrent.Future
 
@@ -26,5 +30,22 @@ class DiscordBot @Inject()(configuration: Configuration, eventListener: DiscordB
             DiscordLogger.info("Monster Truck Bot logging out")
             client.logout()
         }))
+    }
+
+    def getInviteLink(guildId: String, redirectTo: String): String = {
+        val inviteBuilder: BotInviteBuilder = new BotInviteBuilder(client)
+        inviteBuilder.withPermissions(java.util.EnumSet.of[Permissions](
+            Permissions.ADMINISTRATOR,
+            Permissions.KICK,
+            Permissions.BAN,
+            Permissions.READ_MESSAGES,
+            Permissions.SEND_MESSAGES,
+            Permissions.EMBED_LINKS,
+            Permissions.ATTACH_FILES,
+            Permissions.MENTION_EVERYONE,
+            Permissions.VOICE_CONNECT,
+            Permissions.VOICE_SPEAK
+        ))
+        inviteBuilder.build() + "&guild_id=" + guildId + "&response_type=code&redirect_uri=" + URLEncoder.encode(redirectTo, StandardCharsets.UTF_8.name())
     }
 }
